requirements.txt
# Python dependencies for Trading Algorithm scripts
requests>=2.25.0
pandas>=1.3.0
python-dotenv>=0.19.0 